refactor: SubmissionStateDto location
