+++ Source/Tests/Kysect.Shreks.Tests/Application/UpdateSubmissionStateHandlerTest.cs
@@ -1,6 +1,6 @@
 using FluentAssertions;
 using Kysect.Shreks.Application.Abstractions.Submissions.Commands;
-using Kysect.Shreks.Application.Dto.Users;
+using Kysect.Shreks.Application.Dto.Study;
 using Kysect.Shreks.Application.Handlers.Submissions;
 using Microsoft.EntityFrameworkCore;
 using Xunit;
